Referral Generator Complete Project
Allows user to fill in the potential lead data in an interactive web form and then autofills the required pdf referral form and consent forms. Locates the concerned CCU and autopopulates this info and Combines these two into one single pdf and sends the pdf as an email attachement to required emails.